games/requirements_game.txt
# Requirements for Web Interface and Evaluation
# Minimal dependencies for running web UI and evaluation framework

# Web framework
fastapi>=0.116.0
uvicorn>=0.35.0
pydantic>=2.0.0
websockets>=15.0.0

# Configuration management (Hydra-based)
hydra-core>=1.3.0
omegaconf>=2.3.0
PyYAML>=6.0.0

# Agent framework
agentscope==1.0.8

# Logging
loguru>=0.7.0

# Numerical computing
numpy>=1.24.0

# OpenAI API client (for model calls)
openai>=1.0.0

# HTTP client (for API calls)
httpx>=0.28.0
requests>=2.32.0

# Async support
nest-asyncio>=1.6.0

# Diplomacy support
diplomacy>=1.1.2

# Token counter (HuggingFaceTokenCounter)
transformers==4.53.2

# Template engine
Jinja2==3.1.6
